+++ guru-vynil-master-suite/Source/PluginProcessor.h
@@ -40,6 +40,7 @@ public:
 
 private:
     VinylMasterDSP dsp;
+    juce::AudioBuffer<float> dryBuffer;
 
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(GuruVynilMasterSuiteAudioProcessor)
 };